docs/_code_samples/quickstart/py/single-asset-vault/requirements.txt
# xrpl-py==4.1.0
xrpl-py @ git+https://github.com/ckeshava/xrpl-py.git@sav
ttkbootstrap==1.12.0
